Separate SaveButton logic to each *Admin component
